Refactor Boundary component
