+++ src/frontend/src/components/dashboard/DashboardWidget.tsx
@@ -2,7 +2,7 @@ import { t } from '@lingui/core/macro';
 import { ActionIcon, Box, Group, Overlay, Paper, Tooltip } from '@mantine/core';
 import { IconX } from '@tabler/icons-react';
 
-import { Boundary } from '../Boundary';
+import { Boundary } from '@lib/components/Boundary';
 
 import type { ModelType } from '@lib/index';
 import type { JSX } from 'react';
