+++ src/frontend/src/components/panels/PanelGroup.tsx
@@ -32,6 +32,7 @@ import {
   useParams
 } from 'react-router-dom';
 
+import { Boundary } from '@lib/components/Boundary';
 import type { ModelType } from '@lib/enums/ModelType';
 import { identifierString } from '@lib/functions/Conversion';
 import { cancelEvent } from '@lib/functions/Events';
@@ -43,7 +44,6 @@ import { generateUrl } from '../../functions/urls';
 import { usePluginPanels } from '../../hooks/UsePluginPanels';
 import { useLocalState } from '../../states/LocalState';
 import { vars } from '../../theme';
-import { Boundary } from '../Boundary';
 import { StylishText } from '../items/StylishText';
 import type { PanelGroupType, PanelType } from '../panels/Panel';
 import * as classes from './PanelGroup.css';
