+++ src/frontend/src/components/plugins/RemoteComponent.tsx
@@ -3,6 +3,7 @@ import { Alert, MantineProvider, Stack, Text } from '@mantine/core';
 import { IconExclamationCircle } from '@tabler/icons-react';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
+import { Boundary } from '@lib/components/Boundary';
 import { identifierString } from '@lib/functions/Conversion';
 import type { InvenTreePluginContext } from '@lib/types/Plugins';
 import { type Root, createRoot } from 'react-dom/client';
@@ -10,7 +11,6 @@ import { api, queryClient } from '../../App';
 import { ApiProvider } from '../../contexts/ApiContext';
 import { LanguageContext } from '../../contexts/LanguageContext';
 import { useLocalState } from '../../states/LocalState';
-import { Boundary } from '../Boundary';
 import { findExternalPluginFunction } from './PluginSource';
 
 /**
