+++ src/frontend/src/components/settings/SettingItem.tsx
@@ -13,13 +13,13 @@ import {
 import { IconEdit } from '@tabler/icons-react';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
+import { Boundary } from '@lib/components/Boundary';
 import { ModelInformationDict } from '@lib/enums/ModelInformation';
 import { ModelType } from '@lib/enums/ModelType';
 import { apiUrl } from '@lib/functions/Api';
 import type { Setting } from '@lib/types/Settings';
 import { api } from '../../App';
 import { vars } from '../../theme';
-import { Boundary } from '../Boundary';
 import { RenderInstance } from '../render/Instance';
 
 type ConfirmResult = {
